feat(absences): demi-journée de congé payé, du solde au bulletin
Demande RH du 04/09 : poser 0,5 jour de CP et le voir partout.

Données : jours_payes passe en numeric (corrige à la racine le 22P02 du
03/09 — paiement par pas de 0,5, planché au demi-jour) ; nouvelle colonne
demi_journees jsonb {"2026-09-14": "matin"} sur absence_requests (jour
absent = jour plein, aucun backfill). Le jour de planning converti porte
quotite_absence=0.5 + demi_journee, clés serveur (SERVER_OWNED) donc non
falsifiables et purgées à l'annulation.

Solde : quotite_demi_journees() pondère l'unique point de décompte
(count_absence_days_taken) — validés, pending et engagements suivent.

Paie : l'analyzer émet l'événement CP d'une demi-journée même quand des
heures sont pointées ce jour-là (l'autre demi-journée est travaillée), et
la normalisation serveur ne force plus le réel à 0 sur ces jours ;
calcul_brut/forfait somment les quotités (retenue et indemnité 0,5 jour,
arbitrage maintien/dixième inchangé).

Au passage, bug latent corrigé : conges_payes entre dans
TYPES_SIGNIFICATIFS_A_ZERO_HEURE — un CP validé via le module absences
(projeté à 0 h) était SUPPRIMÉ par l'agrégation : aucune ligne CP au
bulletin, jamais d'arbitrage 1/10e. Neutre sur le brut sous maintien.

Interface : choix Journée / ½ Matin / ½ Après-midi jour par jour dans la
saisie de congé, quotités affichées (0,5 / 2,5 j) sur les pages Absences,
la file manager, l'historique salarié et les compteurs calendrier.

+++ backend/app/modules/payroll/engine/calcul_conges.py
@@ -39,7 +39,7 @@ def _lire_majoration_hs(contexte: ContextePaie) -> float:
 
 
 def calculer_indemnite_conges(
-    contexte: ContextePaie, nombre_jours_conges: int, salaire_horaire_base: float
+    contexte: ContextePaie, nombre_jours_conges: float, salaire_horaire_base: float
 ) -> Dict[str, Any]:
     """
     Calcule l'indemnité de congés payés en comparant les deux méthodes
